refactor(ocr): decompose OcrService into AnthropicOcrClient/Extractors + parsing util
Behavior-preserving, pure (0 DB/tx/JE). Facade keeps 8-method surface + 1-arg ctor +
delegates; internally constructs sub-services (spec + module + 3 consumers untouched).
Lazy Anthropic client singleton + confidence-retry loop kept intact; Thai prompt
strings + JSON-sanitize logic moved VERBATIM. tsc 0, ocr tests green, lint 0.
